requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
line-bot-sdk==3.8.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
requests==2.31.0
ollama==0.1.6
